backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2

# FIXED
pydantic>=2.9,<3.0
requests>=2.32.5

python-dotenv==1.0.1
httpx==0.27.0
scikit-learn==1.4.1.post1
sentence-transformers==2.6.1
groq>=0.4.1

# File Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==5.1.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
python-multipart==0.0.6
